backend/requirements.txt
fastapi[standard]==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pyjwt==2.10.1
bcrypt==4.2.1
pydantic-settings==2.7.1
playwright==1.49.1
apscheduler==3.10.4
beautifulsoup4==4.12.3
lxml==5.3.0
httpx==0.28.1
